requirements.txt
google-generativeai>=0.3.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
PyPDF2>=3.0.0
tqdm>=4.66.1
Flask>=3.0.0 